fix api delete student

+++ src/modules/study-group/services/study-group.service.ts
@@ -1,4 +1,10 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { In, Repository } from 'typeorm';
 import { PaginationDto } from 'src/common/dtos';
@@ -535,26 +541,39 @@ export class StudyGroupService {
       return { ...item.student };
     });
   }
+
   async deleteStudentFromStudyGroup(studyGroupId: number, studentCode: string) {
-    const studyGroup = await this.studyGroupRepository.findOne({
-      where: { id: studyGroupId },
-      relations: ['group_students', 'group_students.student'],
-    });
-
-    if (!studyGroup) {
-      throw new Error('Nhóm học phần không tồn tại');
+    if (!studentCode) {
+      throw new BadRequestException('Mã sinh viên không hợp lệ');
     }
+    try {
+      const studyGroup = await this.studyGroupRepository.findOne({
+        where: { id: studyGroupId },
+        relations: ['group_students', 'group_students.student'],
+      });
 
-    const student = studyGroup.group_students.find(
-      (groupStudent) => groupStudent.student.student_code === studentCode,
-    );
-    if (!student) {
-      throw new Error('Sinh viên không tồn tại trong nhóm học phần này');
+      if (!studyGroup) {
+        throw new NotFoundException('Nhóm học phần không tồn tại');
+      }
+
+      const student = studyGroup.group_students.find(
+        (groupStudent) => groupStudent.student.student_code === studentCode,
+      );
+      if (!student) {
+        throw new NotFoundException(
+          'Sinh viên không tồn tại trong nhóm học phần này',
+        );
+      }
+      await this.groupStudentRepository.remove(student);
+      return {
+        success: true,
+        message: 'Xóa sinh viên khỏi nhóm thành công',
+      };
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        throw error;
+      }
+      throw new BadRequestException('Lỗi khi xóa sinh viên khỏi nhóm học phần');
     }
-    await this.groupStudentRepository.remove(student);
-    return {
-      success: true,
-      message: 'Xóa sinh viên khỏi nhóm thành công',
-    };
   }
 }
